Fix #110562: Crash animating node property affecting socket visibility
For various reasons, the animation system can't properly update the node tree
so that the socket availability caused by changing node enum properties
propagates completely. So animating node properties that affect
socket visibility to change isn't possible without issues like crashes.
Unfortunately that wasn't disallowed before.

---

We could choose to disallow animating just the enum values that cause
crashes (sockets that cause visibility changes), or we could disallow animation
on all sockets that might be changed to cause visibility updates in the future
so users don't depend on it when it might not work. Both approaches are
"conservative" in different ways, I'm not sure which to take.