Fix #114515: Cycles: Numerical precision issues in triangle light sampling
The refactor in 97d9bbbc97f8 changed the way q is computed in the spherical
triangle sampling code. While the new approach is more efficient and saves
a few operations, it introduces numerical precision issues for skinny/small
(spherical) triangles.

Therefore, this change moves the computation of q back to the method from
the paper, while keeping the more efficient solid angle computation.