UI: Extend tip description of Grid, UV Sphere and Ico Sphere
Currently when beginner is adding new sphere primitives,
the description is not informative. It is especially visible
for not native English speakers.

With this commit I have added basic information what every sphere is,
and basic usage of it. After using this commit, it is much more clear,
when to use each sphere and what is the main differences.

i