Fix #110751: Clear all strokes prior to baking lineart
This is a attempt to fix line art crashing from baking while there are
already strokes in the gpencil object that's not generated by lineart.

Side effect: If you bake the second time in a different frame range, the
previous bake is going to be cleared.

This crash is probably related to the change to using job system
(instead of a simple loop), and during
`BKE_scene_graph_update_for_newframe()` call, blender crashes trying to
duplicate some existing strokes.