Fix to Graph Editor - precision mode doesn't activate #119387
The graph editor defaults to a snapped mode, to the previous changes to not enabling mouse precision broke the expected behaviour on the graph editor.

The fix was to only apply the changes to 3d views and image views, where snapping is not the expected default behaviour, but something that is selected more regularly.