Fix various crashes with the catalog tree, allow refreshing
The popup would close and free data before the catalog tree handling
ran. Generally interactions with the catalog tree shouldn't close the
popup, and supporting refreshing of the popup (so that the layout can
change and the popup size too) is important to support. So now the popup
stays open until a new brush is selected, which I think is the desired
behavior.