Fix #119514: EEVEE-Next: Mixed resolution with orthographic projection
When using mixed resolution rendering with an orthographic projection
the screen was not rendered correctly. Most of the scene was not there.
The reason was that the view boundaries where extracted from the winmat
correctly, but when re-composing the matrix it assumed to be viewed in
perspective.

This is fixed by composing an orthographic winmat when it was sourced
from a orthographic winmat.