EEVEE-Next: Add light spread support
This is just a proof of concept.

This only supports disk lights. Computing for square and
elipses might prove to be more difficult.

This has energy problem as the cycles version does
focuses the light to keep the light power constant.
This version only masks the lighting. Which still
gives correct lighting intensity around the light
but it decays too fast.
However, this raises question about the bounds of the
light: Should they be increased to account for the
focussing? Should we also cull them better in the
light culling phase?

Also the two `acos` are quite expensive. We might want
to have a more simpler expression.