BLF: simplify text shader texel_fetch
Avoid very costly integer division/modulo as well as two texelFetch
calls separated by a branch. We know that font texture width
is power of two, so we can replace division/modulo with a shift
and a mask, that is set from the calling code via glyph_tex_size
uniform.

Analyzing the shader with Mali Offline Compiler for Mali-G76 arch:
- Stack spilling: 724 -> 692 bytes
- Total cycles, arithmetic: 119.5 -> 33.2
- Total cycles, load/store: 161.0 -> 119.0
- Total cycles, texture: 88.0 -> 44.0