EEVEE-Next: Refactor light data packing
This allows to use unions on the C++ side and safe type
casting on the GPU side.

The type casting functions are statically verified at
compile time in C++.

This PR doesn't change the size of the light struct
but removes the need of packing floats in the `object_mat`.
The matrix will be changed to a `float4x3` in another PR
and will reduce the struct by 16 bytes.

This remove the need for the light parameters macros and
reveals the padding members that could be used for each type.

After this, all accesses to area and sun type data in the
shaders should be done using
`LightAreaData light_area_data_get(LightData light)`
and
`LightSunData light_sun_data_get(LightData light)`
respectively.

Note that this is a simple passthrough for Metal since it
supports `union`.