EEVEE-Next: Add operator to convert a world volume to mesh
EEVEE-Next world volume are infinite like Cycles. EEVEE-Classic world volumes
ended at the clip_end of the camera/viewport. This can lead to confusion as
it would render different then expected.

This PR adds an operator to convert a world volume to a mesh volume.
The mesh volume is bound to the active scene camera.

**Why an operator?**

As this alters the content of the scene we want the artist to be in control of
the conversion.

**TODO**

Add the operator to the Outliner and as an object operator in the 3d viewport.

matches Cycles more closer.
Scenes that
are created with EEVEE will