Fix #119754: Edit mode modifier deform data copy crash
Caused by ddcfc46ee6575bc90846e6e0a76ae72cd7313337.

That commit assumed that the cage mesh always had the deformed
position and normal data. But that isn't true, as shown in the example
in the report. To fix the bug, simplify things, and make the goal clearer,
just copy the struct directly.