UI: rephrase Compact with Milliseconds user pref for Timecode Style
The user preference Timecode Style has a "Compact with Milliseconds"
item, but they are not actually milliseconds. For example at 25 fps,
frame 42 will be at 00:01+17, or 00:01.68, which is just seconds with
the decimal part. Actual milliseconds would be 00:01+680ms or
something, which would be cumbersome.

This commit rephrases the option to "Compact with Decimals" instead of
"Milliseconds".