Fix #119692: Account for mesh lights with a light tree root index of 0
When using light linking with the light tree, mesh lights can have
a root node index of 0. This causes issue with the current code
resulting in the wrong PDF being computed during forward path tracing.

So we adjust the code to allow root node indexs of 0.
This was worked on by @Alaska, @Sergey, and @weizhen