GPU: Gracefully handle compressed textures on Metal
Metal does not support compressed textures, but it does not gracefully
handle the cases where compressed textures are used. To fix this, we
extend the capabilities module to check for support of compressed
textures, assert when creating compressed textures when the backend
doesn't support it, and finally fallback to uncompressed textures in the
IMB GPU utilities, which is the only place where compressed textures are
created.