Docs: Python: Refactor the config file out of 'sphinx_doc_gen.py'
This was done originally because we need access to some information that is avalable from `bpy` however, we only need a few pieces of information.

Instead, I have refactored the code to only write this information to 'BLENDER_INFO.py' which is imported from 'conf.py'.

The goal here is to make it simpler to adjust the configuration.

I have provided fallbacks incase the import fails.

This change also moves the config generation from `rna2sphinx` to the main function.