Mesh: Improve and simplify modifier evaluation logic
Fixes #119938

Make use of GeometrySet and implicit sharing to significantly
simplify the handling of mesh evaluation in the modifier stack.

TODO: Proper description, testing

One thing to check is whether it's actually okay to store mesh_input
as a read-only mesh in the evaluated geometry set. I would hope so,
but we may be required to store it separately and add some complexity
back, I'm not sure.