Fix #120058: Crash on undo rename in edit mode
Drop the check for the return of `BKE_id_new_name_validate`, as it doesn't indicate if the name changed. More correct would be to replace it with a check for if the new name is different than the old, but I don't see any neat way to do that from down here.

Other check might also be replaced with an assert, but I'm less sure about that?