Geometry: Use implicit sharing for deformed positions
Avoid copying the positions array into the evaluated edit hints array
that's used to support editing with deformed positions when there is
a topology-changing procedural operation.

This adds more complexity than I hoped, mainly because we don't have
an easy way to combine optionally shared data with RAII. I hope we can
find some abstraction to clean this up in the future, avoiding the need
for manually defining the rule of five for these edit hints structs.