Fix #120175: File Output node has wrong BW output
The File Output node produces wrong output when the output is set to BW.
This is because the image saving logic assumes float buffers always have
4 channels and ignores the image buffer channels member. So for single
channel images, the code weighted sum neighbouring pixels and asigned
them as a single pixel, producing bleeding at best and crashes at worst.

To fix this, we skip Color to BW conversion if the image is already BW.
Ideally, we should fix the functions that assume 4 channels for float
buffers, but this should be done ourside of this patch since it is
intended to be backported. The wrong functions are IMB_color_to_bw and
IMB_saturation.